requirements.txt
# Core dependencies for Bot Factory
aiogram==3.21.0
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.42
python-dotenv==1.1.1
structlog==25.4.0
aiohttp==3.12.15
pydantic==2.11.7
pydantic-settings==2.10.1

# Database migrations  
alembic==1.16.4
psycopg2-binary==2.9.10

# Additional dependencies for stability
typing-extensions==4.14.1
aiofiles==24.1.0
